move new classes to inner namesapce
