+++ Assets/Characters/Enemy/EnemyScript.cs
@@ -1,4 +1,4 @@
-using Core;
+using Core.CharacterTypes;
 
 public class EnemyScript : Enemy
 {
