+++ Assets/Characters/Player/Player.cs
@@ -1,4 +1,4 @@
-using Core;
+using Core.CharacterTypes;
 
 public class Player : Playable
 {
